requirements.txt
langchain-openai
langchain_community
langchain-huggingface
python-dotenv
pypdf